refactor: imports to use services directory
Updated all imports from '@/lib/supabase' and related submodules to '@/services/supabase' and its submodules across admin, API, and auth pages. This change improves code organization by consolidating Supabase-related logic under a dedicated services directory.
