+++ src/app/api/products/images/route.ts
@@ -1,11 +1,11 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { addProductImage, setImageAsPrimary } from '@/lib/supabase/products/product.service';
+import { addProductImage, setImageAsPrimary } from '@/services/supabase/products/product.service';
 
 export async function POST(request: NextRequest): Promise<NextResponse> {
   try {
     // Parse the request body
     const requestBody = await request.json();
-    
+
     // Validate required fields
     if (!requestBody.product_id) {
       return NextResponse.json(
@@ -13,41 +13,41 @@ export async function POST(request: NextRequest): Promise<NextResponse> {
         { status: 400 }
       );
     }
-    
+
     if (!requestBody.image_url) {
       return NextResponse.json(
         { error: 'Image URL is required' },
         { status: 400 }
       );
     }
-    
+
     // Add the image to the product
     const { data, error } = await addProductImage(
       requestBody.product_id,
       requestBody.image_url,
       requestBody.is_primary || false
     );
-    
+
     if (error) {
       return NextResponse.json(
         { error: `Failed to add product image: ${error.message}` },
         { status: 500 }
       );
     }
-    
+
     // If the image should be primary and it was successfully added
     if (requestBody.is_primary && data) {
       // Since addProductImage doesn't return the image_id directly,
       // we need to query for it or extract it from the response
       // For now, we'll skip setting it as primary directly in this step
-      
+
       // In a real implementation, either:
       // 1. Modify addProductImage to return the image_id, or
       // 2. Query for the newly added image to get its ID
-      
+
       console.log('New image added, might need to set as primary:', data);
     }
-    
+
     return NextResponse.json({
       success: true,
       message: 'Product image added successfully',
